Rename connectAdvanced to connect
