+++ src/exports.ts
@@ -1,9 +1,6 @@
 import Provider from './components/Provider'
 import type { ProviderProps } from './components/Provider'
-import connect, {
-  ConnectProps,
-  ConnectedProps,
-} from './components/connectAdvanced'
+import connect, { ConnectProps, ConnectedProps } from './components/connect'
 import type {
   SelectorFactory,
   Selector,
